Move scanline and Barycentric methods to separate classes
